Add newer generic/abstract targets
"image" can be thought of as creating a base for a virtualization
runtime (e.g. docker, podman). This can be later converted to images if
desired.

Deploy/dismantle are sorta like the equivalent for install/uninstall but
for infrastructure. It wouldn't make as much sense to say, install a VM
and uninstall the VM. This applies to containers well.

Deploying some form of virtual infrastructure is straight forward but
dismantling does the reverse of deploying but undoing changes needed to
get that virtual infrastructure in place (e.g. removing container,
virtual bridge, reverting hypervisor settings, so on).