Add MONITOR tag for redundant executables
While each common makefile could be responsible for all the executables
it needs to run its rules. This would part partially undoing the efforts
to consolidate makefile "code" and reduce the management overhead.

That, and check_executables would be performing redundant checks to
ensure an executable can be found on the PATH from a shell.