Add findings for executables var in projects
For one example, see the following project commit:
https://github.com/cavcrosby/general-purpose-scripts/commit/25e1c6c1

The executables variable may have been made simply expanded to allow for
more determistic behavior. This probably is not needed if the project
does not rely on any common makefile executables. Hence those can
remain as recursive variables. Some testing was done and I did not
notice a difference when switching between the two.,