Install makefiles system wide vs per SUDO_USER
Mainly because it would be more "correct" to install these makefiles
system wide. For instance, a ansible_user installs these makefiles but
intends for another user to make use of them entirely. While it might be
possible to have the package manager install the package for a
particular user, this would not be expected intended behavior from a
package manager.