Add mechanism for deploying containers in CI
This is one of many changes needed to be able to deploy containers in a
CI/CD pipeline. For this repo and change, it was done as I will not be
able to be at the console to type in a password when the
docker-test-deploy target runs in the pipeline.