Add the docker-publish rule
Instead of duplicating the same Docker repos into different variables
for different rules, I chose to take the target images (e.g. test vs
release) and put them into one variable. The intended way they are
meant to be used will be up to the targets to handle.

docker push does have a --all-tags option but this would push every
tag known for a local Docker repo to the registry. Which isn't what I
want currently as this would be inconsistent behavior when using the
IMAGE_RELEASE_BUILD make configuration to differentiate what images to
work with.