Correct the docker-test-deploy-dismantle target
Technically, in GNU make, if a target exists in more than one rule with
a recipe then this will only cause a warning to occur and make will
continue. However, this is a violation to other implementations of make
and the POSIX standard.

So while technically nothing will break, this is to mainly inline with
other written rules.