Revert most of 48fe6883
Originally the intent was not to use pyenv as part of the infrastructure
for CI/CD builds on my Poseidon Kubernetes cluster. However, I remember
poetry being unable to use the Python virtual environment it would
create. Hence, I ended installing and using pyenv but solely for
management of Python virtual environments.